userService의 create 메서드 테스트

+++ src/test/java/com/ll/exam/sbb/QuestionRepositoryTests.java
@@ -113,7 +113,7 @@ public class QuestionRepositoryTests {
     @Test
     void findAllPageable() {
         // (re commit) Pageble : 한 페이지에 몇개의 아이템이 나와야 하는지 + 현재 몇 페이지인지)
-        Pageable pageable = PageRequest.of(0, lastSampleDataId);
+        Pageable pageable = PageRequest.of(0, (int)lastSampleDataId);
         Page<Question> page = questionRepository.findAll(pageable);
 
         assertThat(page.getTotalPages()).isEqualTo(1);
